完善UI创建表

+++ DBMS/DBMS.cpp
@@ -7,6 +7,7 @@ DBMS::DBMS(QWidget *parent)
 	this->setFixedSize(this->width(), this->height());
 
 	tableMenu = new QMenu();
+	isCreateTable = false;
 
 	connect(ui.refresh, SIGNAL(triggered()), this, SLOT(sysAction()));
 	connect(ui.commit, SIGNAL(triggered()), this, SLOT(sysAction()));
@@ -48,8 +49,8 @@ void DBMS::disConnAll()
 	disconnect(ui.newTable, SIGNAL(triggered()), this, SLOT(tableAction()));
 	disconnect(ui.deleteTable, SIGNAL(triggered()), this, SLOT(tableAction()));
 	disconnect(ui.editTable, SIGNAL(triggered()), this, SLOT(tableAction()));
-	disconnect(ui.insertTable, SIGNAL(triggered()), this, SLOT(tableAction()));
-	disconnect(ui.saveTabel, SIGNAL(triggered()), this, SLOT(tableAction()));
+	disconnect(ui.insertTable, SIGNAL(triggered()), this, SLOT(insertTableRow()));
+	disconnect(ui.saveTabel, SIGNAL(triggered()), this, SLOT(saveTable()));
 
 	disconnect(ui.newField, SIGNAL(triggered()), this, SLOT(fieldAction()));
 	disconnect(ui.deleteField, SIGNAL(triggered()), this, SLOT(fieldAction()));
@@ -94,11 +95,14 @@ void DBMS::treeClicked(QTreeWidgetItem *item, int col)
 	}
 	else if (item->childCount() == 0)
 	{
+		if (parent->parent() == NULL)
+			return;
 		string d = parent->parent()->text(0).toStdString();
 		string t = parent->text(0).toStdString();
 		string f = item->text(0).toStdString();
 		vector<vector<string>> field = cp.getField(d, t, f);
 
+		ui.table->clear();
 		ui.table->setRowCount(field.size());
 		ui.table->setColumnCount(3);
 		QStringList header{ QString::fromLocal8Bit("约束名"), QString::fromLocal8Bit("列名"), QString::fromLocal8Bit("约束内容") };
@@ -116,7 +120,8 @@ void DBMS::treeClicked(QTreeWidgetItem *item, int col)
 		string t = item->text(0).toStdString();
 		vector<vector<string>> table = cp.getTableInfo(d, t);
 
-		ui.table->setRowCount(table.size());
+		ui.table->clear();
+		ui.table->setRowCount(table.size() - 1);
 		ui.table->setColumnCount(table[0].size());
 		QStringList header;
 		for (auto s : table[0])
@@ -158,7 +163,7 @@ void DBMS::contextMenuEvent(QContextMenuEvent *event)
 			connect(ui.deleteDb, SIGNAL(triggered()), this, SLOT(dbAction()));
 			connect(ui.newTable, SIGNAL(triggered()), this, SLOT(tableAction()));
 		}
-		else if (treeItem->childCount() == 0)
+		else if (treeItem->childCount() == 0 && parent->parent() != NULL)
 		{
 			tableMenu->addAction(ui.newField);
 			tableMenu->addAction(ui.deleteField);
@@ -180,8 +185,8 @@ void DBMS::contextMenuEvent(QContextMenuEvent *event)
 			connect(ui.newTable, SIGNAL(triggered()), this, SLOT(tableAction()));
 			connect(ui.deleteTable, SIGNAL(triggered()), this, SLOT(tableAction()));
 			connect(ui.editTable, SIGNAL(triggered()), this, SLOT(tableAction()));
-			connect(ui.insertTable, SIGNAL(triggered()), this, SLOT(tableAction()));
-			connect(ui.saveTabel, SIGNAL(triggered()), this, SLOT(tableAction()));
+			connect(ui.insertTable, SIGNAL(triggered()), this, SLOT(insertTableRow()));
+			connect(ui.saveTabel, SIGNAL(triggered()), this, SLOT(saveTable()));
 			connect(ui.newField, SIGNAL(triggered()), this, SLOT(fieldAction()));
 		}
 		// 菜单出现的位置为当前鼠标的位置
@@ -264,20 +269,70 @@ void DBMS::dbAction()
 	if (s == "newDb")
 	{
 		QTreeWidgetItem *db = new QTreeWidgetItem(ui.tree, QStringList() << QString::fromLocal8Bit("输入库名"));
+		db->setFlags(db->flags() | Qt::ItemIsEditable);
 		ui.tree->editItem(db);
-		connect(ui.tree, SIGNAL(returnPressed()), this, SLOT(createDb()));
+		connect(ui.tree, SIGNAL(itemChanged(QTreeWidgetItem*, int)), this, SLOT(createDb(QTreeWidgetItem*, int)));
 	}
 	else
 	{
-		string ss = "drop database " + ui.tree->currentItem()->text(0).toStdString() + ";";
-		preSql.push_back(ss);
+		QTreeWidgetItem *db = ui.tree->currentItem();
+		if (db != NULL)
+		{
+			while (db->childCount() > 0)
+			{
+				QTreeWidgetItem *table = db->takeChild(0);
+				while (table->childCount() > 0)
+					delete table->takeChild(0);
+				delete table;
+			}
+			ui.tree->takeTopLevelItem(ui.tree->currentIndex().row());
+			string ss = "drop database " + ui.tree->currentItem()->text(0).toStdString() + ";";
+			delete db;
+			preSql.push_back(ss);
+		}
 	}
 }
 
-void DBMS::createDb()
+/*void CDataTree::closeEditor(QWidget * editor, QAbstractItemDelegate::EndEditHint hint)
 {
-	disconnect(ui.tree, SIGNAL(returnPressed()), this, SLOT(createDb()));
-	string s = "create database " + ui.tree->currentItem()->text(0).toStdString() + ";";
+	QString strOldName = GetOSGObject(this->currentItem())->GetPaiObject()->GetName();
+	QString strNewName = this->currentItem()->text(0);
+
+	CBaseObject * pObject = GetOSGObject(this->currentItem());
+	PaiObject * pParent = pObject->GetPaiObject()->GetParent();
+	QList<PaiObject*> ChildList;
+	pParent->GetChildren(ChildList);
+	int nChildCount = pParent->GetChildrenCount();
+	for (int i = 0; i < nChildCount; i++)
+	{
+		if (ChildList.at(i) == pObject->GetPaiObject())
+		{
+			continue;
+		}
+		QString strName = ChildList.at(i)->GetName();
+		if (strName == strNewName)
+		{
+			this->currentItem()->setText(0, strOldName);
+
+			//editor->setFocus();
+			//this->openPersistentEditor(this->currentItem());
+
+			QMessageBox::warning(NULL, "警告", "名称不能重复！");
+			//this->editItem(this->currentItem());
+			this->setItemSelected(this->currentItem(), true);
+
+			return;
+
+
+		}
+	}
+}*/
+
+void DBMS::createDb(QTreeWidgetItem *item, int col)
+{
+	disconnect(ui.tree, SIGNAL(itemChanged(QTreeWidgetItem*, int)), this, SLOT(createDb(QTreeWidgetItem*, int)));
+	item->setFlags(item->flags() & ~Qt::ItemIsEditable);
+	string s = "create database " + item->text(0).toStdString() + ";";
 	preSql.push_back(s);
 }
 
@@ -285,64 +340,115 @@ void DBMS::tableAction()
 {
 	QString s = sender()->objectName();
 	QTreeWidgetItem *parent = ui.tree->currentItem()->parent();
+	if (parent == NULL)
+		parent = ui.tree->currentItem();
 	string ss = "use database " + parent->text(0).toStdString() + ";";
 	preSql.push_back(ss);
 	if (s == "newTable")
 	{
 		QTreeWidgetItem *table = new QTreeWidgetItem(parent, QStringList() << QString::fromLocal8Bit("输入表名"));
+		table->setFlags(table->flags() | Qt::ItemIsEditable);
+		ui.tree->scrollToItem(table);
 		ui.tree->editItem(table);
-		connect(ui.tree, SIGNAL(returnPressed()), this, SLOT(createTable()));
+		isCreateTable = true;
+		connect(ui.tree, SIGNAL(itemChanged(QTreeWidgetItem*, int)), this, SLOT(createTable(QTreeWidgetItem*, int)));
 	}
-	else
+	else if(s == "dropTable")
 	{
 		ss = "drop table " + ui.tree->currentItem()->text(0).toStdString() + ";";
 		preSql.push_back(ss);
 	}
+	else
+	{
+		QTableWidgetItem *item = ui.table->currentItem();
+		if (item != NULL)
+			ui.table->editItem(item);
+	}
 }
 
-void DBMS::createTable()
+void DBMS::createTable(QTreeWidgetItem *item, int col)
 {
-	disconnect(ui.tree, SIGNAL(returnPressed()), this, SLOT(createTable()));
+	disconnect(ui.tree, SIGNAL(itemChanged(QTreeWidgetItem*, int)), this, SLOT(createTable(QTreeWidgetItem*, int)));
+	item->setFlags(item->flags() & ~Qt::ItemIsEditable);
 
-	ui.table->setRowCount(1);
-	ui.table->setColumnCount(8);
+	ui.table->clear();
+	ui.table->setColumnCount(7);
 	QStringList header{ QString::fromLocal8Bit("列名"),
 						QString::fromLocal8Bit("类型"),
 						QString::fromLocal8Bit("长度"),
 						QString::fromLocal8Bit("主键"),
 						QString::fromLocal8Bit("唯一"),
 						QString::fromLocal8Bit("非空"),
-						QString::fromLocal8Bit("默认值"),
 						QString::fromLocal8Bit("其他约束") };
 	ui.table->setHorizontalHeaderLabels(header);
 	insertTableRow();
+
+	connect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
 }
 
 void DBMS::insertTableRow()
 {
 	ui.table->insertRow(ui.table->rowCount());
 
-	ui.table->setItem(ui.table->rowCount() - 1, 0, new QTableWidgetItem());
-	QStringList typeList{ "integer", "bool", "double", "varchar", "datetime" };
-	QComboBox *type = new QComboBox();
-	type->addItems(typeList);
-	ui.table->setCellWidget(ui.table->rowCount() - 1, 1, type);
-	ui.table->setItem(ui.table->rowCount() - 1, 2, new QTableWidgetItem());
+	if (!isCreateTable)
+		for (int i = 0; i < ui.table->columnCount(); i++)
+			ui.table->setItem(ui.table->rowCount() - 1, i, new QTableWidgetItem());
+	else
+	{
+		ui.table->setItem(ui.table->rowCount() - 1, 0, new QTableWidgetItem());
+		QStringList typeList{ "integer", "bool", "double", "varchar", "datetime" };
+		QComboBox *type = new QComboBox();
+		type->addItems(typeList);
+		ui.table->setCellWidget(ui.table->rowCount() - 1, 1, type);
+		ui.table->setItem(ui.table->rowCount() - 1, 2, new QTableWidgetItem());
 
-	QTableWidgetItem *checkPri = new QTableWidgetItem();
-	checkPri->setCheckState(Qt::Unchecked);
-	ui.table->setItem(ui.table->rowCount() - 1, 3, checkPri);
-	QTableWidgetItem *checkUni = new QTableWidgetItem();
-	checkUni->setCheckState(Qt::Unchecked);
-	ui.table->setItem(ui.table->rowCount() - 1, 4, checkUni);
-	QTableWidgetItem *checkNull = new QTableWidgetItem();
-	checkNull->setCheckState(Qt::Unchecked);
-	ui.table->setItem(ui.table->rowCount() - 1, 5, checkNull);
+		QTableWidgetItem *checkPri = new QTableWidgetItem();
+		checkPri->setCheckState(Qt::Unchecked);
+		ui.table->setItem(ui.table->rowCount() - 1, 3, checkPri);
+		QTableWidgetItem *checkUni = new QTableWidgetItem();
+		checkUni->setCheckState(Qt::Unchecked);
+		ui.table->setItem(ui.table->rowCount() - 1, 4, checkUni);
+		QTableWidgetItem *checkNull = new QTableWidgetItem();
+		checkNull->setCheckState(Qt::Unchecked);
+		ui.table->setItem(ui.table->rowCount() - 1, 5, checkNull);
 
-	ui.table->setItem(ui.table->rowCount() - 1, 6, new QTableWidgetItem());
-	ui.table->setItem(ui.table->rowCount() - 1, 7, new QTableWidgetItem());
+		ui.table->setItem(ui.table->rowCount() - 1, 6, new QTableWidgetItem());
+	}
+}
 
-	//connect();
+void DBMS::consCheck(int row, int col)
+{
+	if (ui.table->item(row, col)->checkState() == Qt::Checked) // 选中
+	{
+		disconnect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+		if (col == 3)
+		{
+			ui.table->item(row, 4)->setFlags(Qt::NoItemFlags);
+			ui.table->item(row, 5)->setFlags(Qt::NoItemFlags);
+		}
+		else if (col == 4 || col == 5)
+			ui.table->item(row, 3)->setFlags(Qt::NoItemFlags);
+		connect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+	}
+	/*else if (ui.table->item(row, 3)->checkState() == Qt::Unchecked) // 未选中
+	{
+		disconnect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+		ui.table->item(row, 4)->setFlags(Qt::ItemIsTristate);
+		ui.table->item(row, 5)->setFlags(Qt::ItemIsTristate);
+		connect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+	}
+	else if (ui.table->item(row, 4)->checkState() == Qt::Unchecked && ui.table->item(row, 5)->checkState() == Qt::Unchecked) // 未选中
+	{
+		disconnect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+		ui.table->item(row, 3)->setFlags(Qt::ItemIsTristate);
+		connect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+	}
+	else if (col == 1)
+	{
+		disconnect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+
+		connect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
+	}*/
 }
 
 void DBMS::saveTable()
@@ -356,7 +462,10 @@ void DBMS::saveTable()
 			return;
 		}
 		s += ui.table->item(i, 0)->text().toStdString() + " ";
-		s += ui.table->item(i, 1)->text().toStdString();
+
+		QComboBox *type = (QComboBox *)ui.table->cellWidget(i, 1);
+		s += type->currentText().toStdString();
+
 		if (ui.table->item(i, 2)->text() != "")
 			s += "(" + ui.table->item(i, 2)->text().toStdString() + ")";
 		if (ui.table->item(i, 3)->checkState() == Qt::Checked)
@@ -369,14 +478,15 @@ void DBMS::saveTable()
 				s += " not null ";
 		}
 		if (ui.table->item(i, 6)->text() != "")
-			s += " default " + ui.table->item(i, 6)->text().toStdString();
-		if (ui.table->item(i, 7)->text() != "")
-			s += " " + ui.table->item(i, 7)->text().toStdString();
+			s += " " + ui.table->item(i, 6)->text().toStdString();
 		s += ",";
 	}
-	s.substr(0, s.size() - 1);
+	s = s.substr(0, s.size() - 1);
 	s += ");";
 	preSql.push_back(s);
+
+	isCreateTable = false;
+	disconnect(ui.table, SIGNAL(cellChanged(int, int)), this, SLOT(consCheck(int, int)));
 }
 
 void DBMS::fieldAction()
